requirements.txt
httpx>=0.27.0
pyyaml>=6.0
curl_cffi>=0.7.0
fake_useragent>=1.5.0
playwright>=1.40.0
playwright-stealth>=1.0.6
